Move optionKeys function to OptionsManager
