+++ src/beautifier.ts
@@ -1,7 +1,7 @@
 import { Language } from "./language";
 import { LanguageManager, LanguageQuery } from "./LanguageManager";
 import { OptionsRegistry } from "./options";
-import { OptionsManager } from "./OptionsManager";
+import { OptionsManager, optionKeys } from "./OptionsManager";
 import { InlineFlagManager } from "./InlineFlagManager";
 import { DependencyDefinition, DependencyManager } from "./DependencyManager";
 import { zipObject, unique } from "./utils";
@@ -599,48 +599,4 @@ export class Unibeautify {
     this.optionsManager.loadOptions(options);
     return this;
   }
-}
-
-export function optionKeys(
-  beautifier: Beautifier,
-  language: Language
-): BeautifierOptionName[] {
-  const beautifierOptions = beautifier.options[language.name];
-  // Transform options
-  if (typeof beautifierOptions === "boolean") {
-    return [];
-  } else if (typeof beautifierOptions === "object") {
-    const options: BeautifierOptionName[] = [];
-    Object.keys(beautifierOptions).forEach(fieldKey => {
-      const op = (<BeautifierLanguageOptionComplex>beautifierOptions)[fieldKey];
-      if (typeof op === "string") {
-        options.push(op);
-      } else if (isOptionTransformSingleFunction(op)) {
-        options.push(fieldKey as BeautifierOptionName);
-      } else if (typeof op === "boolean") {
-        if (op === true) {
-          options.push(fieldKey as BeautifierOptionName);
-        }
-      } else if (isOptionTransform(op)) {
-        options.push(...op[0]);
-      } else {
-        return new Error(
-          `Invalid option "${fieldKey}" with value ${JSON.stringify(op)}.`
-        );
-      }
-    });
-    return options;
-  } else {
-    return [];
-  }
-}
-
-function isOptionTransformSingleFunction(
-  option: any
-): option is BeautifyOptionTransformSingleFunction {
-  return typeof option === "function";
-}
-
-function isOptionTransform(option: any): option is BeautifyOptionTransform {
-  return Array.isArray(option);
 }
